// js/sdks/packages/node-sdk/src/deployment-json-types.ts
// generated by app.quicktype.io

export interface LunaSecDeploymentJSON {
    id:              string;
    project_id:      string;
    organization_id: string;
    deployment_json: DeploymentJSON;
    timestamp:       number;
}

export interface DeploymentJSON {
    name:                   string;
    secrets:                Secret[];
    project_id:             string;
    deployment_id:          string;
    build_id:               string;
    global_handlers:        GlobalHandlers;
    workflow_relationships: any[];
    workflow_states:        WorkflowState[];
}

export interface GlobalHandlers {
}

export interface Secret {
    name: string;
    arn:  string;
}

export interface WorkflowState {
    transitions:            Transitions;
    id:                     string;
    name?:                  string;
    type:                   string;
    bucket_name?:           string;
    cors?:                  boolean;
    language?:              string;
    code?:                  string;
    libraries?:             any[];
    max_execution_name?:    number;
    policies?:              Policy[];
    container?:             Container;
    environment_variables?: EnvironmentVariables;
    arn?:                   string;
    publish?:               boolean;
    api_path?:              string;
    http_method?:           string;
    lambda_proxy?:          string;
    url?:                   string;
}

export interface Container {
    uri: string;
}

export interface EnvironmentVariables {
    LAMBDA_CALLER:            string;
    DOCUMENT_VAULT_S3_BUCKET: string;
    SIGNING_KEYS_ARN:         string;
    CDN_CONFIG:               string;
    SECURE_FRAME_KEYSET_ARN:  string;
    AUTH_PROVIDER_PUBLIC_KEY: string;
    AUTH_CALLBACK_URL:        string;
}

export interface Policy {
    action:   string[];
    resource: string;
}

export interface Transitions {
    if:        any[];
    else:      any[];
    exception: any[];
    then:      any[];
}